Feat: Markets, transactions - Changes: core deps
- Market Creation

Added a market creation page to the `yew-webapp`.
Does basic sanity checks on the market details, then posts the market to Nostr relays.

- Market List

Added a market list page to the `yew-webapp`.
Lists all markets on the network.

- Transactions

Added a transactions page to the `yew-webapp`.
Lists all transactions for the wallet, and links to the transaction details on Mutinynet mempool.

- Core changes

Removes dependency on `nostr` crate for `PredictionMarket` and `Bet` structs, to
integrate in to yew-webapp. Open to suggestions on how to integrate this.