Removes nostr dependency
Removes dependency on `nostr` crate and replaces it with manual
verification.

No new crates are needed as all verification is done using secp256k1
signatures.

This commit also adds a new `PredictionOutcome` struct to the
`market` module, which is used to represent the outcome of a
prediction market.

The `PredictionMarket` struct now generates its own market ID
by creating a Nostr representation of the market and hashing it.

Both of these changes were made to ensure both markets and outcomes can
are verifiable and recontructable as Nostr events.